requirements.txt
openai
crawl4ai @ git+https://github.com/unclecode/crawl4ai.git
groq
twilio
brave-search
datetime